changes: 	- default export changed to named 	- compiled to es5 instead of 2020 	- typings moved to their own files 	- README changed
