+++ src/index.ts
@@ -1,1 +1,1 @@
-module.exports = require("./lib/eth/eth-addresses-observer.js").default;
+export * from "./lib/eth/eth-addresses-observer.js";
