+++ src/lib/eth/eth-addresses-observer.ts
@@ -18,28 +18,31 @@ along with eth-address-observer.  If not, see <https://www.gnu.org/licenses/>.
  * @date 2020
  */
 
-import {
-	IEthAddressesObserver,
-	IEthAddressesObserverConfig,
-	IAddressesObserverConfig,
-	SubscriptionType
-} from "typings";
 import Web3 from "web3";
-import { AddressesObserver } from "../addresses-observer";
+import {
+	AddressesObserver,
+	AddressesObserverConfig
+} from "../addresses-observer";
 import { EthBlocksCollector } from "./eth-blocks-collector";
 import { EthTransactionsCollector } from "./eth-transactions-collector";
 import { EthTransactionsManager } from "./eth-transactions-manager";
 
-export default class EthAddressesObserver
-	extends AddressesObserver
-	implements IEthAddressesObserver {
+export type EthAddressesObserverConfig = Partial<AddressesObserverConfig>;
+
+export type SubscriptionType =
+	| "pending"
+	| "confirmation"
+	| "success"
+	| "dropped";
+
+export default class EthAddressesObserver extends AddressesObserver {
 	ethBlocksCollector: EthBlocksCollector;
 	ethTransactionsCollector: EthTransactionsCollector;
 	ethTransactionsManager: EthTransactionsManager;
 
-	constructor(web3: Web3, config: IEthAddressesObserverConfig = {}) {
+	constructor(web3: Web3, config: EthAddressesObserverConfig = {}) {
 		config.confirmationsRequired = config.confirmationsRequired || 12;
-		super(config as IAddressesObserverConfig);
+		super(config as AddressesObserverConfig);
 
 		this.ethBlocksCollector = new EthBlocksCollector(
 			web3,
