+++ tsconfig.json
@@ -1,10 +1,12 @@
 {
 	"compilerOptions": {
-		"lib": ["ES2020"],
+		// "lib": ["ES2020"],
 		"module": "CommonJS",
-		"target": "ES2020",
+		"target": "ES5",
 		"outDir": "dist/",
 		"rootDir": "src/",
+		"declaration": true,
+		"declarationDir": "typings",
 		"allowJs": true,
 		"esModuleInterop": true,
 		"importHelpers": true,
