+++ typings/lib/eth/eth-transactions-manager.d.ts
@@ -1,0 +1,17 @@
+/**
+ * @file eth-transactions-manager.ts
+ * @author Vitaly Snitovets <v.snitovets@gmail.com>
+ * @date 2020
+ */
+/// <reference types="node" />
+import Web3 from "web3";
+import { EventEmitter } from "events";
+export declare class EthTransactionsManager extends EventEmitter {
+    private readonly web3;
+    private readonly confirmationsRequired;
+    private transactions;
+    constructor(web3: Web3, confirmationsRequired: number);
+    add(transactionHash: string): void;
+    process(latestBlockNumber: number): void;
+    private remove;
+}
